downloader/requirements.txt
boto3~=1.17.67
requests~=2.25.1
python-dateutil~=2.8.1